+++ __tests__/validate.test.js
@@ -6,6 +6,7 @@ import {
   validatePurchaseUnit,
   validateWinningNumberisNumeric,
   validateWinningNumberDuplicate,
+  validateBonusNumberUnique,
 } from "../src/validate";
 
 test("구입급액이 1,000원 단위가 아닐 경우 예외를 발생시킨다.", () => {
@@ -58,4 +59,13 @@ test("당첨 번호가 서로 중복되는 경우 예외를 발생시킨다.", (
   expect(() => {
     validateWinningNumberDuplicate(numbers);
   }).toThrow("당첨 번호는 중복되지 않아야 합니다");
+});
+
+test("당첨 번호와 보너스 번호가 중복되는 경우 예외를 발생시킨다.", () => {
+  const winningNumbers = [1, 2, 3, 4, 5, 6];
+  const bonusNumber = 1;
+
+  expect(() => validateBonusNumberUnique(winningNumbers, bonusNumber)).toThrow(
+    "보너스 번호는 당첨 번호와 중복되면 안됩니다."
+  );
 });
